Change veh from smallint to int

+++ db-server/init/01_init_schema.sql
@@ -92,7 +92,7 @@ CREATE TABLE tl_bus (
   desi            text,
   dir             smallint,
   oper            smallint          NOT NULL,
-  veh             smallint          NOT NULL,
+  veh             integer           NOT NULL,
   tst             timestamptz       NOT NULL,
   tsi             integer,
   spd             real,
